+++ tools/libexec/perl/TTP/MongoDB.pm
@@ -76,7 +76,7 @@ sub _connect {
 		msgVerbose( __PACKAGE__."::_connect() already connected" );
 
 	} else {
-		my( $account, $passwd ) = $self->_getCredentials();
+		my( $account, $passwd ) = $self->_getCredentials( undef, { node => $self->hostingNode() });
 		if( length $account && length $passwd ){
 			my $server = $self->connectionString() || 'localhost:27017';
 			#print STDERR "server $server\n";