+++ ChangeLog.md
@@ -16,6 +16,7 @@
     - TTP::commandExec() now honors <ENVIRONMENT> macro (todo #177)
     - daemon.pl status publishes its telemetry to MQTT and as text (todo #55, #57)
     - daemon.pl sql has a --force option to skip all safety checks
+    - DBMS::getCredentials() searches for the credentials in the hosting node
 
 ### 4.31.0